py/requirements.txt
# Minimum packages for azure functions
azure-functions
azure-functions-worker
protobuf==3.6.1
six==1.12.0

# Additional packages
cowpy==1.1.0
numpy==1.16.4
onnxruntime==0.5.0
Pillow==6.1.0
requests==2.22.0
